requirements.txt
python-telegram-bot==20.7
matplotlib==3.7.2
Pillow==10.0.1
psycopg2-binary==2.9.7
numpy==1.26.0